+++ src/jquery.fancytree.edit.js
@@ -291,13 +291,16 @@ $.ui.fancytree.registerExtension({
 		this.$container.addClass("fancytree-ext-edit");
 	},
 	nodeClick: function(ctx) {
-		if( $.inArray("shift+click", ctx.options.edit.triggerStart) >= 0 ){
+		var eventStr = $.ui.fancytree.eventToString(ctx.originalEvent),
+			triggerStart = ctx.options.edit.triggerStart;
+
+		if( eventStr === "shift+click" && $.inArray("shift+click", triggerStart) >= 0 ){
 			if( ctx.originalEvent.shiftKey ){
 				ctx.node.editStart();
 				return false;
 			}
 		}
-		if( $.inArray("clickActive", ctx.options.edit.triggerStart) >= 0 ){
+		if( eventStr === "click" && $.inArray("clickActive", triggerStart) >= 0 ){
 			// Only when click was inside title text (not aynwhere else in the row)
 			if( ctx.node.isActive() && !ctx.node.isEditing() &&
 				$(ctx.originalEvent.target).hasClass("fancytree-title")